network: split windows/unix specific code into their own C files
Instead of having a huge #ifdef/#else/#endif block with the same
functions defined differently on each branch, move these functions to
network-unix.c and network-windows.c, which are only compiled for their
respective platform.
