+++ network.h
@@ -1,0 +1,50 @@
+/* SPDX-License-Identifier: LGPL-2.1-or-later */
+/*
+ * libiio - Library for interfacing industrial I/O (IIO) devices
+ *
+ * Copyright (C) 2014-2020 Analog Devices, Inc.
+ * Author: Paul Cercueil
+ *         Robin Getz
+ */
+
+#ifndef __IIO_NETWORK_H
+#define __IIO_NETWORK_H
+
+#include "iio-config.h"
+
+#include <stdbool.h>
+
+struct addrinfo;
+
+struct iio_network_io_context {
+	int fd;
+
+	/* Only buffer IO contexts can be cancelled. */
+	bool cancellable;
+	bool cancelled;
+#if defined(_WIN32)
+	void * events[2];
+#elif defined(WITH_NETWORK_EVENTFD)
+	int cancel_fd[1]; /* eventfd */
+#else
+	int cancel_fd[2]; /* pipe */
+#endif
+	unsigned int timeout_ms;
+};
+
+int setup_cancel(struct iio_network_io_context *io_ctx);
+void cleanup_cancel(struct iio_network_io_context *io_ctx);
+void do_cancel(struct iio_network_io_context *io_ctx);
+int wait_cancellable(struct iio_network_io_context *io_ctx, bool read);
+
+int do_create_socket(const struct addrinfo *addrinfo);
+
+int set_blocking_mode(int s, bool blocking);
+int set_socket_timeout(int fd, unsigned int timeout);
+
+int network_get_error(void);
+bool network_should_retry(int err);
+bool network_is_interrupted(int err);
+bool network_connect_in_progress(int err);
+
+#endif /* __IIO_NETWORK_H */
